files-view: use explicit in_destruction boolean
So we were relying on the model to be NULL to not update all
the view widgets and details in the done_loading callback.

However, we need the model alive so in the signal of is-loading
emitted on the done_loading function, external objects can peek
the model, etc. even in destruction time.

So instead of setting the model as NULL on destroy and relying on it,
use a explicit boolean to not update view widgets on destroy.